+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export interface UserSettings {
+    currency: string
+    language: string
+    region: string
+}
